Feat: #59 - map 함수 분리
