+++ src/components/main/MapSection.tsx
@@ -1,16 +1,12 @@
+import { IMarker } from '@@types/map';
 import COLOR from '@constants/colors';
 import FONT from '@constants/fonts';
+import { changeMarker, initTmap } from '@utils/map';
+import axios from 'axios';
 import { useEffect, useRef, useState } from 'react';
 import styled from 'styled-components';
-import { changeMarker, initTmap } from '@utils/maps';
 
-interface IMarker {
-  id: number;
-  lat: string;
-  lng: string;
-  name: string;
-  type: 'artGallery' | 'museum' | 'exhibition';
-}
+const CENTER = { LAT: '37.53084364186228', LNG: '127.081908811749' };
 
 const MapSection = () => {
   const mapRef = useRef<HTMLDivElement | null>(null);
@@ -18,29 +14,29 @@ const MapSection = () => {
   const [markerList, setMarkerList] = useState<IMarker[]>([
     {
       id: 0,
-      lat: '37.519892712436906',
-      lng: '127.02810900563199',
+      latitude: '37.519892712436906',
+      longitude: '127.02810900563199',
       name: '고궁1',
       type: 'artGallery'
     },
     {
       id: 1,
-      lat: '37.53288934463672',
-      lng: '127.11971717230388',
+      latitude: '37.53288934463672',
+      longitude: '127.11971717230388',
       name: '고궁2',
       type: 'artGallery'
     },
     {
       id: 2,
-      lat: '37.52127761904626',
-      lng: '127.13346617572014',
+      latitude: '37.52127761904626',
+      longitude: '127.13346617572014',
       name: '국립고궁박물관',
       type: 'museum'
     },
     {
       id: 3,
-      lat: '37.5591696189164',
-      lng: '127.07389565460413',
+      latitude: '37.5591696189164',
+      longitude: '127.07389565460413',
       name: '국립현대미술관',
       type: 'exhibition'
     }
@@ -81,7 +77,16 @@ const MapSection = () => {
   }, [tags]);
 
   useEffect(() => {
-    initTmap(markerList).then((markers: any) => {
+    // axios
+    // .get('/api/center')
+    // .then((res) => [res.data.data.latitude, res.data.data.longitude])
+    // .then((res) =>
+    //   initTmap(markerList, res[0], res[1]).then((markers: any) => {
+    //     markersRef.current = markers;
+    //   })
+    // );
+
+    initTmap(markerList, CENTER.LAT, CENTER.LNG).then((markers: any) => {
       markersRef.current = markers;
     });
   }, []);
